refactor(MagicNumber): 매직 넘버 파일 내 상수를 직관적으로 파악하도록 파일 분리
