+++ src/constants/CarConstants.js
@@ -1,0 +1,3 @@
+export const DEFAULT_POSITION = 0;
+
+export const UNIT_POSITION = 1;
