--- src/constants/MagicNumber.js
@@ -1,23 +1,0 @@
-export const DELIMITER = ',';
-
-export const MAXIMUM_COUNT = 1000000;
-
-export const MIN_RANDOM_NUMBER = 0;
-
-export const MAX_RANDOM_NUMBER = 9;
-
-export const MOVE_NUMBER = 4;
-
-export const DEFAULT_POSITION = 0;
-
-export const UNIT_POSITION = 1;
-
-export const RESULT_MARK = '-';
-
-export const WINNER_DELIMITER = ', ';
-
-export const MIN_CAR_NUMBER = 2;
-
-export const MAX_CAR_LENGTH = 5;
-
-export const MIN_TRY_COUNT = 1;
