+++ src/constants/PrinterConstants.js
@@ -1,0 +1,3 @@
+export const RESULT_MARK = '-';
+
+export const WINNER_DELIMITER = ', ';
