+++ src/constants/RacingConstants.js
@@ -1,0 +1,5 @@
+export const MIN_RANDOM_NUMBER = 0;
+
+export const MAX_RANDOM_NUMBER = 9;
+
+export const MOVE_NUMBER = 4;
