+++ src/constants/ValidatorConstants.js
@@ -1,0 +1,9 @@
+export const NAME_DELIMITER = ',';
+
+export const MIN_TRY_COUNT = 1;
+
+export const MAX_TRY_COUNT = 1_000_000;
+
+export const MIN_CAR_NUMBER = 2;
+
+export const MAX_CAR_LENGTH = 5;
