+++ src/utils/splitInput.js
@@ -1,7 +1,7 @@
-import { DELIMITER } from '../constants/MagicNumber.js';
+import { NAME_DELIMITER } from '../constants/ValidatorConstants.js';
 
 function splitInput(input) {
-  return input.split(DELIMITER).map((element) => element.trim());
+  return input.split(NAME_DELIMITER).map((element) => element.trim());
 }
 
 export default splitInput;
